# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iGO - API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/client/orders": {
            "get": {
                "tags": [
                    "Client: Orders"
                ],
                "summary": "Show all client order",
                "description": "Display a list of all client orders ",
                "operationId": "showOrders",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/client/order/inprogress": {
            "get": {
                "tags": [
                    "Client: Orders"
                ],
                "summary": "Show all client orders in progress ",
                "description": "Display a list of all client orders in progress (status id between 3 and 5)",
                "operationId": "showInProgressOrders",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/client/order/checkout": {
            "get": {
                "tags": [
                    "Client: Orders"
                ],
                "summary": "Display order data",
                "description": "Show order details - Products, delivery, schedule, tax and payment data",
                "operationId": "showOrdersDetails",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Client: Orders"
                ],
                "summary": "Checkout order",
                "description": "Update order data and if sent any paramter",
                "operationId": "checkout",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "deliver_at": {
                                        "type": "string",
                                        "example": "2022-12-21 19:44"
                                    },
                                    "tax_name": {
                                        "type": "string",
                                        "example": "Fulano da Silva"
                                    },
                                    "tax_number": {
                                        "type": "integer",
                                        "example": "300 322 984"
                                    },
                                    "campaign_code": {
                                        "type": "string",
                                        "example": "FERIAS2021"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/client/order/submit": {
            "post": {
                "tags": [
                    "Client: Orders"
                ],
                "summary": "Submit order",
                "description": "Change order status from Open to Submitted",
                "operationId": "submitOrder",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/client/order/{id}": {
            "get": {
                "tags": [
                    "Client: Orders"
                ],
                "summary": "Get the specified order",
                "description": "Return data of the specified order",
                "operationId": "getOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}